+++ src/pages/FieldNotes.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+;
+const FieldNotes = () => {
+    return (
+        <div>FieldNotes</div>
+    );
+}
+
+export default FieldNotes